blockchain/web3-tournament/requirements.txt
web3
hexbytes
eth-tester
py-evm
Django
djangorestframework
daphne
adrf
py-solc-x
cbor2
psycopg2-binary
